Script to show ec-policies version deployed
We sometimes need to merge things only when a certain policy update
lands on Konflux prod. This is supposed to provide an easy way to
find out what version of ec-policies is being used in Konflux.

For convenience it shows the commit of the sha that it thinks should
be deployed.