Bash script to show latest build details
I often do the podman run command for the Red Hat build to see what
the latest released version is. This might save some time if the
command falls out of my bash history.

It could also be a handy reference for those wondering where to get
ec-cli container images.