Revert "Resolve TA image reference at runtime"
This reverts commit f229b1301aef244a4428e575b12eacbb530c04c5.

This was breaking ci in the build-definitions repo.

There, the `latest` release of this is pulled and used, which means that
whenever a new trusted artifacts bundle is available, suddenly CI fails
for *all* changes until those bundle references are updated and merged.

We should rely on renovate to bump those references, like other bundle
references so we can manage it in a controlled fashion.